ai-service/pyproject.toml
[tool.poetry]
name = "ai-service"
version = "1.0.0"
description = "Samsara Alert AI Service - Procesa alertas usando OpenAI GPT-4o con Google ADK"
authors = ["SAM Technologies"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.115.0,<1.119.0"
uvicorn = {version = ">=0.34.0,<1.0.0", extras = ["standard"]}
pydantic-settings = "^2.1.0"
httpx = ">=0.28.1,<1.0.0"
python-dotenv = "^1.0.0"
google-adk = "^1.18.0"
litellm = "^1.80.0"
langfuse = "^2.0.0"
samsara-api = "^4.1.0"
twilio = "^9.0.0"
redis = "^5.0.0"
structlog = "^24.0.0"
pandas = "^2.2.0"
numpy = "^2.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.12.0"
respx = "^0.21.1"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

